chore: update issue templates (#798)
<!-- markdownlint-disable MD041 -->
#### What this PR does / why we need it
Add project management ISSUE_TEMPLATES.

**Note**
This means that this repository won't inherit the centrally managed
ISSUE_TEMPLATES. So users will not be able to create `bug` or `feature`
on ocm project but would have to do this on each particular repository.

#### Which issue(s) this PR is related to
<!--
Usage: `Related to #<issue number>`, or `Related to (paste link of
issue)`.
-->

Signed-off-by: Fabian Burth <fabian.burth@sap.com>